% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infected_neighbors.R
\name{find_infected_neighbors}
\alias{find_infected_neighbors}
\title{Function to find for the point infection_matrix[i,j] the neighbors that were infected 
from all the 8 edges from a point in the matrix
For point (x, y) the points (x + 1, y + 1), (x + 1, y), (x + 1, y + 1),
(x, y  + 1), (x, y + 1), (x + 1, y + 1), (x + 1, y), and (x + 1, y + 1) are considered neighbors.}
\usage{
find_infected_neighbors(infection_matrix, i, j)
}
\arguments{
\item{infection_matrix}{: main infection_matrix that maps an individual and 
its current state state (unburnt, on fire or burnt out)}

\item{i}{: current row in the  infection_matrix to detect neighbors}

\item{j}{: current column in the  infection_matrix to detect neighbors}
}
\value{
number_infected: number of infected individuals after checking the 8 neighbors of a point
}
\description{
Function to find for the point infection_matrix[i,j] the neighbors that were infected 
from all the 8 edges from a point in the matrix
For point (x, y) the points (x + 1, y + 1), (x + 1, y), (x + 1, y + 1),
(x, y  + 1), (x, y + 1), (x + 1, y + 1), (x + 1, y), and (x + 1, y + 1) are considered neighbors.
}
\examples{


}
