% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infected_neighbors_apply.R
\name{find_infected_neighbors_apply}
\alias{find_infected_neighbors_apply}
\title{Function to calculate for each point in the infection_matrix the neighbors that were infected 
from all the 8 edges
For point (x, y) the points (x + 1, y + 1), (x + 1, y), (x + 1, y + 1),
(x, y  + 1), (x, y + 1), (x + 1, y + 1), (x + 1, y), and (x + 1, y + 1) are considered neighbors.}
\usage{
find_infected_neighbors_apply(infection_matrix)
}
\arguments{
\item{infection_matrix}{: main infection_matrix that maps an individual and 
its current state state (unburnt, on fire or burnt out)}
}
\value{
number_infected_neigh: vector with the number of infected individuals
for each point in a vector
}
\description{
Function to calculate for each point in the infection_matrix the neighbors that were infected 
from all the 8 edges
For point (x, y) the points (x + 1, y + 1), (x + 1, y), (x + 1, y + 1),
(x, y  + 1), (x, y + 1), (x + 1, y + 1), (x + 1, y), and (x + 1, y + 1) are considered neighbors.
}
\examples{

}
